Interfaces, POCO, Entity Framework Core, SQLite
In this section of the project interfaces were created with only one of
them being implemented. Then a POCO class was created (User.cs) and this
POCO class was Unit Tested in UserTest.cs. More Unit Tests may be added
in the future if time permits.

After this the POCO class was copied into the PlaygroundForEFCoreSQLite
project and a table was created with the user POCO class.